requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
Unipath==1.1
python-dotenv==1.0.0